// types/note.ts
export type NoteTag = `Todo` | `Work` | `Personal` | `Meeting` | `Shopping` 

export interface Note {
  id: string
  title: string
  content: string
  tag: NoteTag
  createdAt: string
  updatedAt: string
}

export interface NewNoteBody {
  title: string
  content: string
  tag: NoteTag
}

export const NoteTags: NoteTag[] = [
  `Todo`,
  `Work`,
  `Personal`,
  `Meeting`,
  `Shopping`,
]